Updated Ubuntu 18.04 gpg-agent kill script to ignore itself

+++ file_roots/setup/ubuntu/gpg_agent.sls
@@ -193,11 +193,14 @@ gpg_agent_ps_kill_script_file_exists:
     - contents: |
         #!/bin/bash
         gpg_active=$(ps -ef | grep -v 'grep' | grep gpg-agent)
+        script_pid=$$
         IFS=$'\n'	# make newlines the only seperator
         if [[ -n "$gpg_active" ]]; then
             for gpg_line in $gpg_active; do
                 ps_gpg_agent=$(echo "$gpg_line" | awk '{print $2}')
-                kill -9 $ps_gpg_agent
+                if [[ "$script_pid" -ne "$ps_gpg_agent" ]]; then
+                    kill -9 $ps_gpg_agent
+                fi
             done
         fi
         unset IFS
